ENH: Merging changes from vxl-upstream-1-2-0 to vxl-upstream-2006-07-20 in Utilities/vxl directory to the ITK development trunk.  Corresponding changes to the rest of ITK have also been made.

+++ Code/Numerics/FEM/itkFEMItpackSparseMatrix.h
@@ -39,7 +39,7 @@ class ItpackSparseMatrix
 public:
 
   /** typedefs from f2c.h  */
-  typedef int       integer;
+  typedef long      integer;
   typedef double    doublereal;
 
   /** Constructor */
@@ -252,7 +252,7 @@ class FEMExceptionItpackSparseMatrixSbagn : public FEMException
 public:
 
   /** typedefs from f2c.h  */
-  typedef int       integer;
+  typedef long      integer;
   typedef double    doublereal;
 
   /**
@@ -280,7 +280,7 @@ class FEMExceptionItpackSparseMatrixSbsij : public FEMException
 {
 public:
   /** typedefs from f2c.h  */
-  typedef int       integer;
+  typedef long      integer;
   typedef double    doublereal;
 
   /**
